Move SessionToken.cs
