--- MinecraftClient/Protocol/SessionCache/CacheType.cs
@@ -1,4 +1,0 @@
-﻿namespace MinecraftClient.Protocol.SessionCache
-{
-    public enum CacheType { None, Memory, Disk };
-}
